Fix CK_PKCS5_PBKD2_PARAMS.PasswordLen type

+++ src/Pkcs11Interop/LowLevelAPI4/MechanismParams/CK_PKCS5_PBKD2_PARAMS.cs
@@ -67,6 +67,6 @@ namespace Net.Pkcs11Interop.LowLevelAPI4.MechanismParams
         /// <summary>
         /// Length in bytes of the password information
         /// </summary>
-        public uint PasswordLen;
+        public IntPtr PasswordLen;
     }
 }
